+++ R/data.R
@@ -20,3 +20,22 @@
 #'
 #'
 "logoTextMD"
+
+
+#' Font Awesome Lookup
+#'
+#' Lookup font awesome icons by groups (icons may be duplicated, i.e. belong to
+#'      more than one group)
+#'
+#' @usage data(fa_lookup)
+#' @format ## `fa_lookup`
+#' A data frame with 2,350 rows and 4 columns:
+#' \describe{
+#'   \item{class}{class name (with leading `fa-` removed), all lowercase, no spaces}
+#'   \item{name}{Name of icon, title case with spaces}
+#'   \item{unicode}{unicode}
+#'   \item{group}{group classification of the icon}
+#'   }
+#'
+#'
+"fa_lookup"